+++ honestcash-v2/src/app/app.module.ts
@@ -15,10 +15,10 @@ import { UserEffects } from './store/user/user.effects';
 import { reducers, metaReducers } from './store/app.states';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
 import { environment } from '../environments/environment';
-import { AuthGuardService as AuthGuard } from './core/services/auth-guard.service';
+import { AuthGuardService as AuthGuard } from './core/guards/auth-guard.service';
 import {
   TokenInterceptor, ErrorInterceptor
-} from './core/services/token.interceptor';
+} from './core/http/token.interceptor';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 
 @NgModule({
