+++ honestcash-v2/src/app/core/auth/authentication.guard.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Router, CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 
-import { Logger } from '../logger.service';
+import { Logger } from '../services/logger.service';
 import { AuthenticationService } from './authentication.service';
 
 const log = new Logger('AuthenticationGuard');
