+++ honestcash-v2/src/app/core/http/http-cache.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpResponse } from '@angular/common/http';
 import { each } from 'lodash';
 
-import { Logger } from '../logger.service';
+import { Logger } from '../services/logger.service';
 
 const log = new Logger('HttpCacheService');
 const cachePersistenceKey = 'httpCache';
