+++ honestcash-v2/src/app/core/http/token.interceptor.ts
@@ -6,7 +6,7 @@ import {
 import { Observable, throwError } from 'rxjs';
 import { catchError } from 'rxjs/operators';
 import { Router } from '@angular/router';
-import { AuthService } from './auth.service';
+import { AuthService } from '../services/auth.service';
 
 
 @Injectable()
