+++ honestcash-v2/src/app/core/index.ts
@@ -8,4 +8,4 @@ export * from './http/api-prefix.interceptor';
 export * from './http/cache.interceptor';
 export * from './http/error-handler.interceptor';
 export * from './route-reusable-strategy';
-export * from './logger.service';
+export * from './services/logger.service';
